fix: address PR review feedback across v3net and config editor
- Fix mnemonic grid overflow: switch TUI display to 3x8 layout (54 chars fits 60-wide box)
- v3net-bootstrap: fail fast if keystore was just created (refuse accidental identity)
- Hub wizard: check saveAll() error before showing success message
- createHubMessageAreas: update existing map after append to prevent duplicate tags
- Hub tests: check keystore.Load errors instead of discarding them
- Split keystore_test.go into keystore_test.go + mnemonic_test.go (300-line guideline)
- Export: refuse to overwrite existing recovery file
- Keystore save: atomic write via temp file + rename with 0600 permissions
- Identity screen: context-aware help bar, improved vertical centering
